requirements.txt
streamlit
langchain_core
langchain_community
langchain_openai
pytube
youtube-transcript-api

